+++ src/screens/Start/index.jsx
@@ -1,28 +1,40 @@
 import React from 'react';
-import {View, Text, Image, StyleSheet} from 'react-native';
-import StartImg from '@assets/images/banner/StartImg.png';
-//import Kakao from '@/components/icon/Kakao';
-// import {KakaoLogin} from '@/api/KakaoLogin';
-// import kakao from '@assets/images/register/kakao.png';
-import {KakaoLogin} from '@/api/KakaoLogin';
+import {
+  View,
+  Text,
+  Image,
+  StyleSheet,
+  Linking,
+  TouchableOpacity,
+} from 'react-native';
+import BannerImg from '@assets/images/banner/BannerImage.png';
+import kakao from '@assets/images/register/kakao.png';
 
-export default function Start({navigation}) {
+export default function Start() {
+  const kakaoURL = 'http://3.39.254.198:8080/oauth2/authorization/kakao';
+
+  const handleLogin = async () => {
+    try {
+      await Linking.openURL(kakaoURL);
+    } catch (error) {
+      console.error('Failed to login with Kakao:', error);
+      alert('Login Failed', 'An error occurred during Kakao login.');
+    }
+  };
+
   return (
     <View style={styles.container}>
       <View style={styles.textContainer}>
-        <Text style={styles.title}>Bridge</Text>
-        <Text style={styles.subtitle}>더 가까운 사이가 되는 중</Text>
+        <Text style={styles.title}>Familing</Text>
+        <Text style={styles.subtitle}>더 가까운 가족이 되는 중</Text>
       </View>
-      <Image source={StartImg} style={styles.image} />
-      {/* <TouchableOpacity
-        style={styles.button}
-        onPress={navigation.navigate('RegisterScreen')}>
+      <Image source={BannerImg} style={styles.image} />
+      <TouchableOpacity style={styles.button} onPress={handleLogin}>
         <View style={styles.btnContainer}>
           <Image source={kakao} style={styles.kakao} />
           <Text style={styles.buttonText}>카카오로 시작하기</Text>
         </View>
-      </TouchableOpacity> */}
-      <KakaoLogin />
+      </TouchableOpacity>
     </View>
   );
 }
@@ -31,7 +43,7 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     alignItems: 'center',
-    backgroundColor: '#BBDEFB',
+    backgroundColor: '#4D83F4',
   },
   textContainer: {
     marginTop: 164,
@@ -52,14 +64,14 @@ const styles = StyleSheet.create({
     color: '#fff',
   },
   image: {
-    width: 200,
-    height: 200,
+    width: 240,
+    height: 240,
   },
   button: {
-    backgroundColor: '#FFD600',
+    backgroundColor: '#FFBE00',
     paddingVertical: 12,
     paddingHorizontal: 80,
-    borderRadius: 40,
+    borderRadius: 32.27,
     justifyContent: 'center',
     alignItems: 'center',
     width: 312,
@@ -71,9 +83,9 @@ const styles = StyleSheet.create({
     gap: 8,
   },
   buttonText: {
-    fontSize: 18,
-    color: '#000',
-    fontWeight: 'bold',
+    fontSize: 16,
+    color: '#371C09',
+    fontWeight: '800',
   },
   kakao: {
     width: 28,
